fix(Dataspreadsheet): refractor larger event handlers to seperate util
